+++ tests/test_dt.py
@@ -160,9 +160,13 @@ def test_dt_properties(dt0):
                           ltype.bool, ltype.bool, ltype.str)
     assert dt0.stypes == (stype.int8, stype.bool8, stype.bool8, stype.float64,
                           stype.bool8, stype.bool8, stype.str32)
-    assert dt0.internal.alloc_size > 500
-    assert sys.getsizeof(dt0) >= dt0.internal.alloc_size
+    assert sys.getsizeof(dt0) > 500
 
+
+def test_sizeof():
+    DT1 = dt.Frame(A=["foo"])
+    DT2 = dt.Frame(A=["foo" * 1001])
+    assert sys.getsizeof(DT2) - sys.getsizeof(DT1) == 3000
 
 
 def test_internal():